+++ src/ts/Editor.ts
@@ -12,16 +12,21 @@ import { gup } from "./gup";
 import { detectDarkMode } from "./detectDarkMode";
 import { MouseHandler, PBParams, PlotBoilerplate, XYDimension } from "plotboilerplate";
 import { RPGDialogueLogic } from "./RPGDialogueLogic";
-import { IDialogueConfig, IMiniQuestionaire, IMiniQuestionaireWithPosition } from "./interfaces";
+import { IDialogueConfig, IDialogueListener, IMiniQuestionaire, IMiniQuestionaireWithPosition } from "./interfaces";
 import { EditorHelper } from "./editorHelpers";
 import { EditorRenderer } from "./editorRenderer";
 import { TouchHandler } from "./TouchHandler";
 import { FileDrop } from "plotboilerplate/src/cjs/utils/io/FileDrop";
+import { Modal } from "./modal";
 
 export class Editor {
   currentMouseHandler: MouseHandler | null = null;
   currentTouchHandler: TouchHandler | null = null;
+  editorHelpers: EditorHelper;
+  editorRenderer: EditorRenderer;
+  dialogConfig: IDialogueConfig<IMiniQuestionaireWithPosition> | null;
   constructor(dialogueConfigJSONPath: string) {
+    const _self = this;
     console.log("Initialize plotboilerplate");
     // Fetch the GET params
     const GUP = gup();
@@ -59,23 +64,22 @@ export class Editor {
       ) as PBParams
     );
 
-    var dialogConfig: IDialogueConfig<IMiniQuestionaireWithPosition> = null;
     const boxSize: XYDimension = {
       width: 120,
       height: 20
     };
-    const editorHelpers = new EditorHelper(this, pb, boxSize);
-    var editorRenderer = new EditorRenderer(pb, boxSize, editorHelpers, isDarkmode);
+    this.editorHelpers = new EditorHelper(this, pb, boxSize);
+    this.editorRenderer = new EditorRenderer(pb, boxSize, this.editorHelpers, isDarkmode);
 
     // +---------------------------------------------------------------------------------
     // | The render method.
     // +-------------------------------
     pb.config.postDraw = function (draw, fill) {
-      if (!dialogConfig) {
+      if (!_self.dialogConfig) {
         return;
       }
-      editorRenderer.renderBoxes(dialogConfig);
-      editorRenderer.renderConnections(dialogConfig);
+      _self.editorRenderer.renderBoxes(_self.dialogConfig);
+      _self.editorRenderer.renderConnections(_self.dialogConfig);
     };
 
     RPGDialogueLogic.loadConfigFromJSON(dialogueConfigJSONPath).then((config: IDialogueConfig<IMiniQuestionaire>) => {
@@ -85,24 +89,24 @@ export class Editor {
 
     const handleDialogConfigLoaded = (config: IDialogueConfig<IMiniQuestionaire>) => {
       // Check if all graph nodes have positions to render.
-      dialogConfig = editorHelpers.enrichPositions(config);
-      editorHelpers.enrichMetaData(dialogConfig);
-      console.log("Enriched meta data", dialogConfig);
-      editorHelpers.setDialogConfig(dialogConfig);
+      _self.dialogConfig = this.editorHelpers.enrichPositions(config);
+      this.editorHelpers.enrichMetaData(_self.dialogConfig);
+      console.log("Enriched meta data", _self.dialogConfig);
+      this.editorHelpers.setDialogConfig(_self.dialogConfig);
 
       // Ad DnD support for boxes.
       if (this.currentMouseHandler) {
         this.currentMouseHandler.destroy();
         this.currentMouseHandler = null;
       }
-      this.currentMouseHandler = editorHelpers.boxMovehandler(); // dialogConfig);
+      this.currentMouseHandler = this.editorHelpers.boxMovehandler(); // dialogConfig);
 
       // Ad DnD support for boxes.
       if (this.currentTouchHandler) {
         this.currentTouchHandler.destroy();
         this.currentTouchHandler = null;
       }
-      this.currentTouchHandler = new TouchHandler(pb, dialogConfig, editorHelpers);
+      this.currentTouchHandler = new TouchHandler(pb, _self.dialogConfig, this.editorHelpers);
 
       pb.redraw();
     };
@@ -120,12 +124,12 @@ export class Editor {
       document.getElementById("input-upload-file").click();
     };
     document.getElementById("b-import-json").addEventListener("click", importJSON);
-    document.getElementById("input-upload-file").addEventListener("change", (evt: Event) => {
+    document.getElementById("input-upload-file").addEventListener("change", (_evt: Event) => {
       var fileInput = document.getElementById("input-upload-file") as HTMLInputElement;
       if (!fileInput.files || fileInput.files.length === 0) {
         return;
       }
-      console.log("pictureFile", fileInput.files[0]);
+      console.log("inputFile", fileInput.files[0]);
       var reader = new FileReader();
       reader.onload = function () {
         const jsonText = reader.result as string;
@@ -135,8 +139,57 @@ export class Editor {
       reader.readAsText(fileInput.files[0]);
     });
 
-    // +---------------------------------------------------------------------------------
-    // | END Editor
-    // +-------------------------------
+    document.getElementById("b-run-test").addEventListener("click", () => {
+      _self.testCurrentDialogueConfig();
+    });
   }
+
+  /**
+   * Open a modal and test the current dialogue config (runs a RPGDialogueLogic instant).
+   */
+  testCurrentDialogueConfig() {
+    const _self = this;
+    // Create this structure:
+    // <div class="rpg-output">
+    //    <div class="rpg-output-question"></div>
+    //    <ul class="rpg-output-options"></ul>
+    // </div>
+
+    const outputContainer = document.createElement("div");
+    const outputQuestion = document.createElement("div");
+    const outputOptions = document.createElement("ul");
+    outputContainer.classList.add("rpg-output");
+    outputQuestion.classList.add("rpg-output-question");
+    outputOptions.classList.add("rpg-output-options");
+    outputContainer.appendChild(outputQuestion);
+    outputContainer.appendChild(outputOptions);
+
+    const dialogueListener: IDialogueListener<IMiniQuestionaireWithPosition> = (
+      dialogueConfig: IDialogueConfig<IMiniQuestionaireWithPosition>,
+      nextNodeName: string,
+      oldNodeName: string,
+      selectedOptionIndex: number
+    ) => {
+      // Highlight current node in the graph editor :)
+      // console.log("nextNodeName", nextNodeName, "oldNodeName", oldNodeName, "selectedOptionIndex", selectedOptionIndex);
+
+      _self.editorHelpers.setHighlightedNode(nextNodeName);
+    };
+
+    const rpgLogic = new RPGDialogueLogic(this.dialogConfig, false);
+    rpgLogic.addDialogueChangeListener(dialogueListener);
+    const alternateStartNodeName: string | null = this.editorHelpers.selectedNodeName;
+    this.editorHelpers.setSelectedNode(null, null);
+    rpgLogic.beginConversation(outputQuestion, outputOptions, alternateStartNodeName);
+
+    this.editorHelpers.domHelper.modal.setTitle("Test");
+    this.editorHelpers.domHelper.modal.setBody(outputContainer);
+    this.editorHelpers.domHelper.modal.setFooter("");
+    this.editorHelpers.domHelper.modal.setActions([Modal.ACTION_CLOSE]);
+    this.editorHelpers.domHelper.modal.open();
+  }
+
+  // +---------------------------------------------------------------------------------
+  // | END Editor
+  // +-------------------------------
 }